Revert "Revert "[IndVars] Replace PHIs if loop exits on 1st iteration""
This reverts commit 0657f6c22de07b5acbcc5c1e131e880b5e0f8fee.

The patch was reverted on incorrect claim that this patch may break LCSSA form
when the loop is not in a simplify form. All IndVars' transform insure that
the loop is in simplify and LCSSA form, so if it wasn't broken before this
transform, it will also not be broken after it.